changed column names for the code-first models

+++ RequestModels/AddProductAndCategoriesRequestModel.cs
@@ -3,9 +3,9 @@ namespace Zadanie10.RequestModels;
 public class AddProductAndCategoriesRequestModel
 {
     public string ProductName { get; set; }
-    public decimal ProductWeight { get; set; }
-    public decimal ProductWidth { get; set; }
-    public decimal ProductHeight { get; set; }
-    public decimal ProductDepth { get; set; }
+    public double ProductWeight { get; set; }
+    public double ProductWidth { get; set; }
+    public double ProductHeight { get; set; }
+    public double ProductDepth { get; set; }
     public List<int> ProductCategories { get; set; }
 }
